+++ src/App.jsx
@@ -5,88 +5,11 @@ import Timer from "./components/Timer";
 import WeatherStat from "./components/WeatherStat";
 import ThemeSwitch from "./components/ThemeSwitch";
 // Goltis data
-import { goltis } from "./goltis";
-
-const timers = [
-  {
-    id: 1,
-    time: 2,
-    isPlaying: false,
-  },
-  {
-    id: 2,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 3,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 4,
-    time: 3,
-    isPlaying: false,
-  },
-  {
-    id: 5,
-    time: 10,
-    isPlaying: false,
-  },
-  {
-    id: 6,
-    time: 2,
-    isPlaying: false,
-  },
-  {
-    id: 7,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 8,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 9,
-    time: 3,
-    isPlaying: false,
-  },
-  {
-    id: 10,
-    time: 10,
-    isPlaying: false,
-  },
-  {
-    id: 11,
-    time: 2,
-    isPlaying: false,
-  },
-  {
-    id: 12,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 13,
-    time: 4,
-    isPlaying: false,
-  },
-  {
-    id: 14,
-    time: 3,
-    isPlaying: false,
-  },
-  {
-    id: 15,
-    time: 10,
-    isPlaying: false,
-  },
-];
+import { goltis } from "./data/goltis";
+import { timers } from "./data/timers";
 
 function App() {
-  const [timersState, setTimersState] = useState(goltis);
+  const [timersState, setTimersState] = useState(timers);
   const [isPlaying, setIsPlaying] = useState(false);
   const [isFinished, setIsFinished] = useState(false);
   //
