+++ src/data/timers.ts
@@ -1,0 +1,77 @@
+export const timers = [
+  {
+    id: 1,
+    time: 2,
+    isPlaying: false,
+  },
+  {
+    id: 2,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 3,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 4,
+    time: 3,
+    isPlaying: false,
+  },
+  {
+    id: 5,
+    time: 10,
+    isPlaying: false,
+  },
+  {
+    id: 6,
+    time: 2,
+    isPlaying: false,
+  },
+  {
+    id: 7,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 8,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 9,
+    time: 3,
+    isPlaying: false,
+  },
+  {
+    id: 10,
+    time: 10,
+    isPlaying: false,
+  },
+  {
+    id: 11,
+    time: 2,
+    isPlaying: false,
+  },
+  {
+    id: 12,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 13,
+    time: 4,
+    isPlaying: false,
+  },
+  {
+    id: 14,
+    time: 3,
+    isPlaying: false,
+  },
+  {
+    id: 15,
+    time: 10,
+    isPlaying: false,
+  },
+];
